5.0.1rel code sync 20241022 update001 Signed-off-by heguangyao <heguangyao1@h-partners.com>

+++ frameworks/kits/js/common/picture_ndk/picture_native.cpp
@@ -446,7 +446,7 @@ Image_ErrorCode OH_AuxiliaryPictureInfo_SetRowStride(OH_AuxiliaryPictureInfo *in
     if (info == nullptr || !info->GetInnerAuxiliaryPictureInfo()) {
         return IMAGE_BAD_PARAMETER;
     }
-    info->GetInnerAuxiliaryPictureInfo()->rowStride = rowStride;
+    info->GetInnerAuxiliaryPictureInfo()->rowStride = static_cast<int32_t>(rowStride);
     return  IMAGE_SUCCESS;
 }
